DPP-11 #DONE tests for service1 added
