+++ service1/application/models.py
@@ -1,0 +1,5 @@
+from application import db
+
+class Winners(db.Model):
+	raceno = db.Column(db.Integer, primary_key=True)
+	name = db.Column(db.String(30), nullable=False)
