--- service1/tests/test_service1
@@ -1,1 +1,0 @@
-import unittests
